Move code in test_add_function into files for easier debugging
