+++ tests/interop/test_add_function.out
@@ -1,0 +1,2 @@
+Hello 7 from JS!
+ok
