+++ tests/interop/test_add_function_post.js
@@ -1,0 +1,4 @@
+var newFuncPtr = Runtime.addFunction(function(num) {
+    Module.print('Hello ' + num + ' from JS!');
+});
+Module.callMain([newFuncPtr.toString()]);
